Move output under runner
