+++ telepresence/main.py
@@ -20,7 +20,7 @@ import sys
 from telepresence import connect, mount, outbound, proxy, remote_env
 from telepresence.runner import wait_for_exit, Runner
 from telepresence.cli import parse_args, crash_reporting
-from telepresence.output import Output
+from telepresence.runner.output import Output
 from telepresence.startup import KubeInfo, final_checks
 from telepresence.usage_tracking import call_scout
